+++ src/otx/core/types/task.py
@@ -35,7 +35,6 @@ class OTXTaskType(str, Enum):
 
     # Visual Promting Tasks.
     VISUAL_PROMPTING = "VISUAL_PROMPTING"
-    ZERO_SHOT_VISUAL_PROMPTING = "ZERO_SHOT_VISUAL_PROMPTING"
 
 
 class OTXTrainType(str, Enum):